refactor: add new play_type routes, rename old routes files
