+++ crates/servers/nfl_server/src/handlers/mod.rs
@@ -1,3 +1,5 @@
-pub mod game_clock;
+pub mod game_clock_handlers;
+pub mod play_type_handlers;
 
-pub use game_clock::{GameClockHandlers, GameClockMigrationHandler};
+pub use game_clock_handlers::{GameClockHandlers, GameClockMigrationHandler};
+pub use play_type_handlers::{PlayTypeHandlers, PlayTypeMigrationHandler};
