+++ crates/servers/nfl_server/src/routes/game_clock_routes.rs
