+++ crates/servers/nfl_server/src/routes/play_type_routes.rs
